refactor: trim action_items, hide memberships, rename requests, split reports
Phase-2 cleanup pass following the architecture audit. Four independent
changes, all gated (typecheck + lint + build + 806 unit tests pass):

1. action_items: trim to the lead-followup worker only.
   The Inbox UI that surfaced action_items is gone, so the createActionItem/
   resolveActionItems writes in the estimate/job/intake/SMS/booking flows were
   dead — every call site's comment referenced the deleted Inbox. Removed those
   writers and the lib/action-items module. The action_items TABLE and the
   lead-followup worker (which owns its own inline insert/dedup) are unchanged,
   so stale-lead follow-ups still work.

2. memberships: hide the maintenance-plans UI.
   Memberships are paused indefinitely. Added a maintenance-plans/layout.tsx
   redirect guard (MEMBERSHIPS_ENABLED=false) so every /app/maintenance-plans/*
   route is unreachable. Pages, API routes, and DB tables are retained for a
   future revival; only the UI surface is hidden.

3. routes: standardize /app/booking-requests/[id] -> /app/requests/[id].
   "Requests" is the product term. Moved the detail subtree, shared
   request-guidance, and its test into requests/; deleted the redirect stub and
   the unused BookingRequestActions component; repointed all UI links. The
   /api/v1/booking-requests API is kept (matches the booking_requests table).

4. reports: split the ~1000-line page.tsx (now ~115 lines).
   Extracted format.ts (helpers), queries.ts (row types + loadReportData), and
   section components (Financial, Technician, Operations, PricingHealth,
   EstimateMargins). Pure refactor — same queries, same rendered sections.
