--- apps/web/app/app/booking-requests/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function BookingRequestsPage() {
-  redirect("/app/requests");
-}
